Much nicer integration of chruby for zsh
chruby wants to do everything exactly the same
for bash and zsh. zsh supports running a command
for a new shell and on each directory change, but
bash doesn’t, so chruby uses the lowest common
denominator functionality and checks the directory
BEFORE each command. This is mostly fine, but is
a tiny bit inefficient and also means you might
get a warning “no such ruby” based on the
directory you’re leaving.

There’s no good way to fix this for bash. But I
don’t care about bash! So I’m stealing the useful
code from chruby and hooking it as a callback
on dirctory change for zsh. Much nicer!