Fix for crashing vim on OSX in tmux
It has been crashing once very 5-10 write operations. I think this is
just because the system version is outdated (7.3, compiled 5+ years ago)
and is therefore mixing the bugfix for this.  MacVim from Homebrew is
much more up-to-date.

See https://groups.google.com/forum/#!topic/vim_dev/jMdy2Pffh_k
for discussion of the crash (at least, I *think* this is the same
issue)